nixos: rebuild less derivations on nixpkgs/config hash changes.
Introduce a top-level file nixos-version.json, and an option to have nixos-version
read this file from /run/current-system instead, of hard-coding all the values in the script.
When configured as such, this avoids rebuilding every derivation that depends on the system path
every time that the nixpkgs or configuration revision hash changes.
For my config, this change reduced the amount of derivations that need to be
rebuild after a trivial config change from 42 to 9.