anki: switch to using qt5
Fixes https://github.com/NixOS/nixpkgs/issues/248357

I'm unsure if the issue is with nixpkgs pyqt6 packaging, or if the issue
is an upstream one.

Either way, we're able to get working preferences by switching back to
qt5, and upstream supports both qt5 and qt6 still, so this seems like a
reasonable path to take.

We can tackle getting qt6 working if it's still broken whenever upstream
or nixpkgs drops qt5 support.